requirements.txt
pydantic>=2.7
SQLAlchemy>=2.0
pandas>=2.2
numpy>=1.26
python-dotenv>=1.0
PyYAML>=6.0
tldextract>=5.1
beautifulsoup4>=4.12
regex>=2024.5.15
vaderSentiment>=3.3.2
streamlit>=1.36
google-generativeai>=0.7

